moved Category and nested data picker to one directory
