+++ src/app/components/component-collection.ts
@@ -44,10 +44,10 @@ import { TerraTimePickerComponent } from './forms/input/time-picker/terra-time-p
 import { TerraNoteComponent } from './note/terra-note.component';
 import { TerraNodeTreeComponent } from './tree/node-tree/terra-node-tree.component';
 import { TerraNodeComponent } from './tree/node-tree/node/terra-node.component';
-import { TerraCategoryPickerComponent } from './category-picker/terra-category-picker.component';
+import { TerraCategoryPickerComponent } from './tree-picker/category-picker/terra-category-picker.component';
 import { TerraDynamicFormComponent } from './forms/dynamic-form/terra-dynamic-form.component';
 import { TerraDynamicSwitchComponent } from './forms/dynamic-form/dynamic-switch/terra-dynamic-switch.component';
-import { TerraNestedDataPickerComponent } from './nested-data-picker/terra-nested-data-picker.component';
+import { TerraNestedDataPickerComponent } from './tree-picker/nested-data-picker/terra-nested-data-picker.component';
 import { TerraTwoColumnsContainerComponent } from './layouts/column-container/two-columns/terra-two-columns-container.component';
 import { TerraFormComponent } from './forms/form/terra-form.component';
 import { TerraFormContainerComponent } from './forms/form/form-container/terra-form-container.component';
