+++ src/app/components/tree-picker/category-picker/data/category-detail-data.interface.ts
