+++ src/app/components/tree-picker/category-picker/data/category-value.interface.ts
