+++ src/app/components/tree-picker/nested-data-picker/config/nested-data-tree.config.ts
@@ -1,4 +1,4 @@
-import { TerraNodeTreeConfig } from '../../tree/node-tree/data/terra-node-tree.config';
+import { TerraNodeTreeConfig } from '../../../tree/node-tree/data/terra-node-tree.config';
 import { TranslationService } from 'angular-l10n';
 import { Injectable } from '@angular/core';
 import { NestedDataInterface } from '../data/nested-data.interface';
