+++ src/app/components/tree-picker/nested-data-picker/data/nested-detail-data.interface.ts
