+++ src/app/components/tree-picker/nested-data-picker/data/nested-value.interface.ts
@@ -1,4 +1,4 @@
-import { TerraPlacementEnum } from '../../../helpers/enums/terra-placement.enum';
+import { TerraPlacementEnum } from '../../../../helpers/enums/terra-placement.enum';
 
 /**
  * @author chirila-ioan-daniel
