+++ src/app/components/tree-picker/nested-data-picker/terra-nested-data-picker.component.scss
