+++ src/app/components/tree-picker/nested-data-picker/terra-nested-data-picker.component.ts
@@ -9,14 +9,14 @@ import { TranslationService } from 'angular-l10n';
 import { NG_VALUE_ACCESSOR } from '@angular/forms';
 import { NestedDataTreeConfig } from './config/nested-data-tree.config';
 import { NestedDataInterface } from './data/nested-data.interface';
-import { TerraNodeInterface } from '../tree/node-tree/data/terra-node.interface';
+import { TerraNodeInterface } from '../../tree/node-tree/data/terra-node.interface';
 import { isNullOrUndefined } from 'util';
 import { NestedValueInterface } from './data/nested-value.interface';
 import { TerraNestedDataPickerBaseService } from './service/terra-nested-data-picker-base.service';
-import { TerraNodeTreeConfig } from '../../components/tree/node-tree/data/terra-node-tree.config';
+import { TerraNodeTreeConfig } from '../../tree/node-tree/data/terra-node-tree.config';
 import { Observable } from 'rxjs/Observable';
 import { NestedDetailDataInterface } from './data/nested-detail-data.interface';
-import { TerraPagerInterface } from '../pager/data/terra-pager.interface';
+import { TerraPagerInterface } from '../../pager/data/terra-pager.interface';
 
 @Component({
     selector:  'terra-nested-data-picker',
