+++ src/index.ts
@@ -104,19 +104,19 @@ export * from './app/components/forms/dynamic-form/service/terra-form-field-cont
 export * from './app/components/file-browser/terra-base-private-storage.interface';
 export * from './app/components/file-browser/terra-file-browser.service';
 export * from './app/components/file-browser/terra-file-browser.component';
-export * from './app/components/category-picker/terra-category-picker.component';
-export * from './app/components/category-picker/service/terra-category-picker-base.service';
+export * from './app/components/tree-picker/category-picker/terra-category-picker.component';
+export * from './app/components/tree-picker/category-picker/service/terra-category-picker-base.service';
 export * from './app/components/pager/data/terra-pager.data';
 export * from './app/models/terra-key-value.interface';
-export * from './app/components/category-picker/terra-category-picker.component';
-export * from './app/components/category-picker/service/terra-category-picker-base.service';
-export * from './app/components/category-picker/data/category-data.interface';
-export * from './app/components/category-picker/data/category-detail-data.interface';
-export * from './app/components/category-picker/data/category-value.interface';
-export * from './app/components/nested-data-picker/terra-nested-data-picker.component';
-export * from './app/components/nested-data-picker/service/terra-nested-data-picker-base.service';
-export * from './app/components/nested-data-picker/data/nested-data.interface';
-export * from './app/components/nested-data-picker/data/nested-value.interface';
+export * from './app/components/tree-picker/category-picker/terra-category-picker.component';
+export * from './app/components/tree-picker/category-picker/service/terra-category-picker-base.service';
+export * from './app/components/tree-picker/category-picker/data/category-data.interface';
+export * from './app/components/tree-picker/category-picker/data/category-detail-data.interface';
+export * from './app/components/tree-picker/category-picker/data/category-value.interface';
+export * from './app/components/tree-picker/nested-data-picker/terra-nested-data-picker.component';
+export * from './app/components/tree-picker/nested-data-picker/service/terra-nested-data-picker-base.service';
+export * from './app/components/tree-picker/nested-data-picker/data/nested-data.interface';
+export * from './app/components/tree-picker/nested-data-picker/data/nested-value.interface';
 export * from './app/helpers/enums/button-color.enum';
 export * from './app/helpers/enums/terra-placement.enum';
 export * from './app/helpers/regex/terra-regex';
